Parse Huffman info
Next step is to create the table.
We will create a binary tree for each Huffman tree.
Left will be assumed to be the '0' path, and right the '1' path.

If, and only if a node has no children, will that node represent the
termination of a string, and a valid value for the binary code-string.