Fix the IDCT->RGB conversion
It seems to me that the meaning of the X index and Y index is inverted
while referring to RGB plane (as opposed to our initial assumption that
x ==  number of rows, and y == columns).

It also turns out that we don't need to reset the DC value after every
block.

This patch contains these fixes.

Change-Id: I0c0cce61e28c72517fcb3ac3a94f7a017f01a1f7